Set `get_preset` for AssetHubs
