+++ system-parachains/asset-hubs/asset-hub-kusama/Cargo.toml
@@ -13,6 +13,7 @@ codec = { features = ["derive", "max-encoded-len"], workspace = true }
 hex-literal = { workspace = true }
 log = { workspace = true }
 scale-info = { features = ["derive"], workspace = true }
+serde_json = { features = ["alloc"], workspace = true }
 
 # Local
 bp-asset-hub-kusama = { workspace = true }
@@ -252,6 +253,7 @@ std = [
 	"polkadot-runtime-constants/std",
 	"primitive-types/std",
 	"scale-info/std",
+	"serde_json/std",
 	"snowbridge-router-primitives/std",
 	"sp-api/std",
 	"sp-block-builder/std",
