+++ system-parachains/asset-hubs/asset-hub-kusama/src/genesis_config_presets.rs
@@ -14,15 +14,15 @@
 // You should have received a copy of the GNU General Public License
 // along with Polkadot.  If not, see <http://www.gnu.org/licenses/>.
 
-//! Genesis configs presets for the CollectivesPolkadot runtime
+//! Genesis configs presets for the AssetHubKusama runtime
 
 use crate::*;
 use sp_std::vec::Vec;
 use system_parachains_constants::genesis_presets::*;
 
-const COLLECTIVES_POLKADOT_ED: Balance = crate::ExistentialDeposit::get();
+const ASSET_HUB_KUSAMA_ED: Balance = crate::ExistentialDeposit::get();
 
-fn collectives_polkadot_genesis(
+fn asset_hub_kusama_genesis(
 	invulnerables: Vec<(AccountId, AuraId)>,
 	endowed_accounts: Vec<AccountId>,
 	id: ParaId,
@@ -32,7 +32,7 @@ fn collectives_polkadot_genesis(
 			balances: endowed_accounts
 				.iter()
 				.cloned()
-				.map(|k| (k, COLLECTIVES_POLKADOT_ED * 4096 * 4096))
+				.map(|k| (k, ASSET_HUB_KUSAMA_ED * 4096 * 4096))
 				.collect(),
 		},
 		"parachainInfo": ParachainInfoConfig {
@@ -41,7 +41,7 @@ fn collectives_polkadot_genesis(
 		},
 		"collatorSelection": CollatorSelectionConfig {
 			invulnerables: invulnerables.iter().cloned().map(|(acc, _)| acc).collect(),
-			candidacy_bond: COLLECTIVES_POLKADOT_ED * 16,
+			candidacy_bond: ASSET_HUB_KUSAMA_ED * 16,
 			..Default::default()
 		},
 		"session": SessionConfig {
@@ -49,9 +49,9 @@ fn collectives_polkadot_genesis(
 				.into_iter()
 				.map(|(acc, aura)| {
 					(
-						acc.clone(),                             // account id
-						acc,                                     // validator id
-						SessionKeys { aura },			 // session keys
+						acc.clone(),                         // account id
+						acc,                                 // validator id
+						SessionKeys { aura }, 	// session keys
 					)
 				})
 				.collect(),
@@ -64,19 +64,19 @@ fn collectives_polkadot_genesis(
 	})
 }
 
-pub fn collectives_polkadot_local_testnet_genesis(para_id: ParaId) -> serde_json::Value {
-	collectives_polkadot_genesis(invulnerables(), testnet_accounts(), para_id)
+pub fn asset_hub_kusama_local_testnet_genesis(para_id: ParaId) -> serde_json::Value {
+	asset_hub_kusama_genesis(invulnerables(), testnet_accounts(), para_id)
 }
 
-fn collectives_polkadot_development_genesis(para_id: ParaId) -> serde_json::Value {
-	collectives_polkadot_local_testnet_genesis(para_id)
+fn asset_hub_kusama_development_genesis(para_id: ParaId) -> serde_json::Value {
+	asset_hub_kusama_local_testnet_genesis(para_id)
 }
 
 /// Provides the JSON representation of predefined genesis config for given `id`.
 pub fn get_preset(id: &sp_genesis_builder::PresetId) -> Option<sp_std::vec::Vec<u8>> {
 	let patch = match id.try_into() {
-		Ok("development") => collectives_polkadot_development_genesis(1001.into()),
-		Ok("local_testnet") => collectives_polkadot_local_testnet_genesis(1001.into()),
+		Ok("development") => asset_hub_kusama_development_genesis(1000.into()),
+		Ok("local_testnet") => asset_hub_kusama_local_testnet_genesis(1000.into()),
 		_ => return None,
 	};
 	Some(
