+++ system-parachains/asset-hubs/asset-hub-polkadot/Cargo.toml
@@ -13,6 +13,7 @@ codec = { features = ["derive", "max-encoded-len"], workspace = true }
 hex-literal = { optional = true, workspace = true }
 log = { workspace = true }
 scale-info = { features = ["derive"], workspace = true }
+serde_json = { features = ["alloc"], workspace = true }
 
 # Local
 bp-asset-hub-kusama = { workspace = true }
@@ -233,6 +234,7 @@ std = [
 	"polkadot-runtime-constants/std",
 	"primitive-types/std",
 	"scale-info/std",
+	"serde_json/std",
 	"snowbridge-router-primitives/std",
 	"sp-api/std",
 	"sp-block-builder/std",
